style. create package 'common' and exception package move to common package
