+++ src/main/java/com/smile/petpat/common/exception/ErrorResponse.java
@@ -1,8 +1,7 @@
-package com.smile.petpat.exception;
+package com.smile.petpat.common.exception;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.Errors;
 import org.springframework.validation.FieldError;
