+++ src/main/java/com/smile/petpat/common/exception/RestApiExceptionHandler.java
@@ -1,15 +1,11 @@
-package com.smile.petpat.exception;
+package com.smile.petpat.common.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.HttpRequestMethodNotSupportedException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.io.InvalidClassException;
-import java.security.InvalidParameterException;
 
 @RestControllerAdvice
 @Slf4j
